Move `StorageId` to `file-store-ot`.
